refactor: day domain
